requirements.txt
langchain_mcp_adapters
langchain_google_genai
langchain
asyncpg
python-dotenv 
